requirements.txt
Flask==2.3.3
requests==2.31.0
python-docx==0.8.11
reportlab==4.0.4
pymupdf4llm
google-generativeai
pinecone
sentence-transformers
faiss-cpu
# RAG runner dependencies
astrapy==2.0.1 
deprecation==2.1.0 
dnspython==2.7.0 
h2==4.2.0 
hpack==4.1.0 
hyperframe==6.1.0 
pymongo==4.14.0 
toml==0.10.2 
uuid6==2025.0.1
# Thread management dependencies
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.16.0
voyageai==0.1.12
tiktoken==0.5.2

